Fix some nasty bugs, thanks to realMain() tests
There were multiple issues. When we exited the
server gracefully we forgot to send the exitCode 
to the quitChannel. Second, we wrongly re-injected
the exitCode into the quitChannel at the end of 
the realMain() method causing a deadlock. The 
issue was actually found because we foolishly 
reused the channels in a test. So, maybe 3 wrongs
make a right? No, it’s just 3 bugs interacting in
weird, unpredictable ways. Also, some tests failed
when then TOKEN env was set.  So that’s actually 4
issues. But this last one was a blessing as that
triggered the bug hunt as the tests were running
fine in one terminal, while failing or hanging
randomly in another!! The only difference between
these two terminals was the TOKEN env variable :)