Implement deadlettering for RabbitMQ and Channel
RabbitMQ: dead letter messages are delivered to
the logproxy exchange with “deadletter.resource” routing
key. Users are free to add subscribers to this key to do
post processing or redelivery. Logproxy does  not
implement this logic itself.

Channel: rejected messages are currently output on
stdout. This may change in the future.